requirements.txt
streamlit
numpy
pandas
requests
Pillow
opencv-python-headless
scikit-image
imutils
